+++ packages/typespec-go/CHANGELOG.md
@@ -1,6 +1,17 @@
 # Release History
 
-## 0.7.1 (unreleased)
+## 0.8.0 (2025-07-22)
+
+### Breaking Changes
+
+* The `module-version` switch has been removed.
+  * For new modules, the `moduleVersion` constant will have an initial value of `v0.1.0`.
+  * For existing modules, the value of `moduleVersion` is externally maintained.
+
+### Features Added
+
+* The `module` switch now accepts a major version suffix (e.g. `module=mymodule/v2`).
+* Added switch `containing-module` for scenarios where the emitted code is a package within an existing module. No `go.mod` or `version.go` file is emitted for this scenario.
 
 ### Bugs Fixed